+++ src/videopoker/userinterface/UserInterface.java
@@ -11,6 +11,6 @@ public interface  UserInterface extends Runnable{
 	public abstract void displayHand(String[] hand);
 	public abstract void displayAdvice(boolean[] advice);
 	public abstract void displayWin(boolean wins,String handPower, int credit);
-	public void displayStats(int credit,int gain,PowerHashMap<String, Integer> statsMap);
+	public void displayStats(int credit,float gain,PowerHashMap<String, Integer> statsMap);
 	public abstract void displayError(String reason);
 }
